sdks/python/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
click>=8.0.0
rich>=13.0.0
pytest>=7.0.0
black>=23.0.0
